all scripts translated
