+++ scripts/src/app/application.terminals.ts
@@ -1,11 +1,11 @@
-import * as terminal from "./terminal";
+import * as appTerminal from "./terminal";
 import * as viewTerminal from "../view/terminal";
 import * as application from "./application";
 
-const items: terminal.Terminal[] = [];
+const items: appTerminal.Terminal[] = [];
 
-export function create(properties: terminal.Properties, viewFactory: viewTerminal.TerminalFactory): terminal.Terminal {
-	var terminal: terminal.Terminal = new terminal.Terminal(properties, application.context, viewFactory);
+export function create(properties: appTerminal.Properties, viewFactory: viewTerminal.TerminalFactory): appTerminal.Terminal {
+	var terminal: appTerminal.Terminal = new appTerminal.Terminal(properties, application.context, viewFactory);
 	items.push(terminal);
 	return terminal;
 }
