+++ scripts/src/app/terminal.ts
@@ -1,332 +1,356 @@
-/// <reference path="./input.ts" />
-/// <reference path="./modules.ts" />
-/// <reference path="./statements.ts" />
-/// <reference path="./application.ts" />
-/// <reference path="../components/types.ts" />
-/// <reference path="../components/commands.ts" />
-/// <reference path="../components/modules.ts" />
-/// <reference path="../components/registry.ts" />
-/// <reference path="../view/terminal.tsx" />
+import * as input from "./input";
+import * as terminal from "./terminal";
+import * as semantics from "./semantics";
+import * as bus from "./application.bus";
+import * as storage from "./storage";
+import * as statements from "./statements";
+import * as modules from "./modules";
+import * as output from "../view/output";
+import * as baseInput from "../view/input/base";
+import * as view from "../view/view";
+import * as viewTerminal from "../view/terminal";
+import * as renderers from "../view/renderers";
+import * as fugaziInput from "../view/input/fugaziInput";
+import * as app from "./application";
+import * as utils from "../core/utils";
+import * as logger from "../core/logger";
+import * as collections from "../core/types.collections";
+import * as coreTypes from "../core/types";
+import * as net from "../core/net";
+import * as components from "../components/components";
+import * as registry from "../components/registry";
+import * as componentsModules from "../components/modules";
+import * as modulesBuilder from "../components/modules.builder";
+import * as componentsBuilder from "../components/components.builder";
+import * as componentsDescriptor from "../components/components.descriptor";
+import * as types from "../components/types";
+import * as syntax from "../components/syntax";
+import * as commands from "../components/commands";
+import * as commandsHandler from "../components/commands.handler";
+import * as handler from "../components/commands.handler";
+import * as descriptor from "../components/commands.descriptor";
 
-module fugazi.app.terminal {
-	export interface Properties {
-		name: string;
-		title?: string;
-		description?: string;
-		history?: string[];
+
+export interface Properties {
+	name: string;
+	title?: string;
+	description?: string;
+	history?: string[];
+}
+
+export abstract class BaseTerminalContext extends app.BaseContext<app.ApplicationContext> {
+	private moduleToRemoteId: collections.Map<string>;
+
+	constructor(parent: app.ApplicationContext, id?: string) {
+		super(parent, id);
+		this.moduleToRemoteId = new collections.Map<string>();
 	}
 
-	export abstract class BaseTerminalContext extends BaseContext<ApplicationContext> {
-		private moduleToRemoteId: collections.Map<string>;
+	public setRemoteSource(modulePath: components.Path, remoteId: string): void {
+		this.moduleToRemoteId.set(modulePath.toString(), remoteId);
+	}
 
-		constructor(parent: ApplicationContext, id?: string) {
-			super(parent, id);
-			this.moduleToRemoteId = new collections.Map<string>();
+	public getRemoteSource(modulePath: components.Path): string {
+		let id: string = this.moduleToRemoteId.get(modulePath.toString());
+
+		if (!coreTypes.isNull(id) || modulePath.first() == modulePath.last()) {
+			return id;
 		}
 
-		public setRemoteSource(modulePath: components.Path, remoteId: string): void {
-			this.moduleToRemoteId.set(modulePath.toString(), remoteId);
+		return this.getRemoteSource(modulePath.parent());
+	}
+}
+
+export class TerminalContext extends BaseTerminalContext {
+	private terminal: Terminal;
+
+	constructor(terminal: Terminal, parent: app.ApplicationContext) {
+		super(parent);
+		this.terminal = terminal;
+	}
+
+	public getTerminal(): Terminal {
+		return this.terminal;
+	}
+
+	public asRestricted(): RestrictedTerminalContext {
+		let context = new RestrictedTerminalContext(this.getParent(), this.getId());
+		Object.seal(context);
+		return context;
+	}
+
+	public isVariableReference(value: string): boolean {
+		return value.startsWith("$");
+	}
+
+	public hasVariable(name: string): boolean {
+		return !coreTypes.isUndefined(this.terminal.retrieveVariable(name));
+	}
+
+	public storeVariable(name: string, type: types.Type, value: any): app.Variable {
+		return this.terminal.storeVariable(name, type, value);
+	}
+
+	public retrieveVariable(name: string): app.Variable {
+		return this.terminal.retrieveVariable(name);
+	}
+
+	public clearOutput(): void {
+		this.terminal.clearOutput();
+	}
+}
+
+export class RestrictedTerminalContext extends BaseTerminalContext {
+}
+
+export class TerminalCommand extends commands.LocalCommand {
+	constructor() {
+		super();
+	}
+
+	public isRestricted(): boolean {
+		return false;
+	}
+}
+
+type LoadedModule = {
+	module: componentsModules.Module;
+	context: {
+		restricted: modules.ModuleContext,
+		privileged: modules.PrivilegedModuleContext
+	};
+};
+
+export type ContextProvider = {
+	terminal: () => TerminalContext;
+	module: {
+		restricted: (path: components.Path) => modules.ModuleContext;
+		privileged: (path: components.Path) => modules.PrivilegedModuleContext;
+	}
+}
+
+function moduleContextProvider(this: { modules: Map<string, LoadedModule> }, type: "restricted" | "privileged", path: components.Path) {
+	while (path.length > 0) {
+		const key = path.toString();
+
+		if (this.modules.has(key)) {
+			return this.modules.get(key).context[type];
 		}
 
-		public getRemoteSource(modulePath: components.Path): string {
-			let id: string = this.moduleToRemoteId.get(modulePath.toString());
+		path = path.parent();
+	}
 
-			if (!isNull(id) || modulePath.first() == modulePath.last()) {
-				return id;
+	return null;
+}
+
+export class Terminal {
+	private properties: Properties;
+	private view: viewTerminal.TerminalView;
+	private context: TerminalContext;
+	private contextProvider: ContextProvider;
+	private modules: Map<string, LoadedModule>;
+	private variables: collections.Map<app.Variable>;
+
+	constructor(properties: Properties, applicationContext: app.ApplicationContext, viewFactory: viewTerminal.TerminalFactory) {
+		this.properties = properties;
+		this.modules = new Map();
+		this.variables = collections.map<app.Variable>();
+		this.context = new TerminalContext(this, applicationContext);
+		this.contextProvider = {
+			terminal: () => this.context,
+			module: {
+				restricted: moduleContextProvider.bind(this, "restricted"),
+				privileged: moduleContextProvider.bind(this, "privileged")
 			}
+		}
 
-			return this.getRemoteSource(modulePath.parent());
+		this.moduleLoaded(registry.getModule("io.fugazi"));
+
+		viewFactory.createTerminal({
+			name: properties.name,
+			title: properties.title || properties.name,
+			description: properties.description,
+			history: this.properties.history,
+			querier: this.queryForStatements.bind(this),
+			executer: this.executeCommand.bind(this)
+		}).then(this.setView.bind(this));
+	}
+
+	public clearOutput(): void {
+		this.view.clearOutput();
+	}
+
+	public moduleLoaded(module: componentsModules.Module) {
+		const path = module.getPath().toString();
+
+		if (!this.modules.has(path)) {
+			this.modules.set(path, {
+				module,
+				context: {
+					restricted: new modules.ModuleContext(this.context.asRestricted()),
+					privileged: new modules.PrivilegedModuleContext(this.context)
+				}
+			});
 		}
 	}
 
-	export class TerminalContext extends BaseTerminalContext {
-		private terminal: Terminal;
-
-		constructor(terminal: Terminal, parent: ApplicationContext) {
-			super(parent);
-			this.terminal = terminal;
-		}
-
-		public getTerminal(): Terminal {
-			return this.terminal;
-		}
-
-		public asRestricted(): RestrictedTerminalContext {
-			let context = new RestrictedTerminalContext(this.getParent(), this.getId());
-			Object.seal(context);
-			return context;
-		}
-
-		public isVariableReference(value: string): boolean {
-			return value.startsWith("$");
-		}
-
-		public hasVariable(name: string): boolean {
-			return !fugazi.isUndefined(this.terminal.retrieveVariable(name));
-		}
-
-		public storeVariable(name: string, type: components.types.Type, value: any): Variable {
-			return this.terminal.storeVariable(name, type, value);
-		}
-
-		public retrieveVariable(name: string): Variable {
-			return this.terminal.retrieveVariable(name);
-		}
-
-		public clearOutput(): void {
-			this.terminal.clearOutput();
-		}
-	}
-
-	export class RestrictedTerminalContext extends BaseTerminalContext {}
-
-	export class TerminalCommand extends components.commands.LocalCommand {
-		constructor() {
-			super();
-		}
-
-		public isRestricted(): boolean {
-			return false;
-		}
-	}
-
-	type LoadedModule = {
-		module: components.modules.Module;
-		context: {
-			restricted: modules.ModuleContext,
-			privileged: modules.PrivilegedModuleContext
+	public storeVariable(name: string, type: types.Type, value: any): app.Variable {
+		let variable = {
+			name: extractVariableName(name),
+			type: type,
+			value: value
 		};
+
+		this.variables.set(variable.name, variable);
+		return variable;
+	}
+
+	public retrieveVariable(name: string): app.Variable {
+		return this.variables.get(extractVariableName(name));
+	}
+
+	private setView(view: viewTerminal.TerminalView): void {
+		this.view = view;
+	}
+
+	private executeCommand(command: string): commands.ExecutionResult {
+		ga("send", "event", "Commands", "execution - start", command);
+		storage.local.store(this.properties.name, this.properties);
+
+		let result: commands.ExecutionResult = null;
+		try {
+			let session = statements.createStatementsSession(command, this.contextProvider),
+				executableStatement = session.getExecutable();
+
+			if (!coreTypes.isNull(executableStatement)) {
+				result = executableStatement.execute();
+			} else {
+				ga("send", "event", "Commands", "execution - None of the statements are executable", command);
+				throw new coreTypes.Exception("None of the statements are executable");
+			}
+		} catch (e) {
+			const error = typeof e === "string" ? e : (e.message ? e.message : e.toString());
+			ga("send", "event", "Commands", "execution - error: " + error, command);
+			result = new commands.ExecutionResult(registry.getType("any"), false);
+			result.reject(e);
+		}
+
+		return result;
+	}
+
+	private queryForStatements(command: string, position: number): Promise<statements.Statement[]> {
+		let future = new coreTypes.Future<statements.Statement[]>(),
+			session = statements.createStatementsSession(command, this.contextProvider),
+			suggestedStatements = session.getSuggestions(position);
+
+		future.resolve(suggestedStatements);
+
+		return future.asPromise();
+	}
+}
+
+function extractVariableName(input: string): string {
+	return input.startsWith("$") ? input.substring(1) : input;
+}
+
+// init
+bus.register(app.Events.Ready, function (): void {
+	let coreCommands = {
+		name: "io.fugazi.terminal",
+		title: "Terminal Module",
+		constraints: {
+			variable: {
+				title: "Variable name constraint",
+				types: ["string"],
+				params: [],
+				validator: () => {
+					return function (value: string): boolean {
+						return /^\$[^"':\[]\(\)\{}]+$/.test(value);
+					}
+				}
+			}
+		},
+		types: {
+			"variable": {
+				title: "Variable Token",
+				type: "string[variable]"
+			}
+		},
+		commands: {
+			clear: {
+				title: "Clears the output panel",
+				syntax: "clear",
+				returns: "void",
+				componentConstructor: TerminalCommand,
+				handler: function (context: modules.PrivilegedModuleContext): handler.Result {
+					context.getParent().clearOutput();
+					return {
+						status: handler.ResultStatus.Success
+					}
+				}
+			},
+			assign: {
+				title: "Variable assignment command",
+				syntax: [
+					"(name string) = (value any)",
+					"set (name string) = (value any)",
+					"let (name string) = (value any)"
+				],
+				returns: "void",
+				parametersForm: "arguments",
+				componentConstructor: TerminalCommand,
+				handler: function (context: modules.PrivilegedModuleContext, name: string, value: any): handler.Result {
+					let type = context.guessType(value);
+					if (type == null) {
+						return {
+							status: handler.ResultStatus.Failure,
+							error: "can't guess type"
+						}
+					}
+
+					context.getParent().storeVariable(name, type, value);
+					return {
+						status: handler.ResultStatus.Success
+					};
+				}
+			},
+			moduleWorkWith: {
+				title: "Specific module remote source setup",
+				syntax: [
+					"module (moduleName string) works with (source string)",
+				],
+				returns: "ui.message",
+				parametersForm: "arguments",
+				componentConstructor: TerminalCommand,
+				handler: function (context: modules.ModuleContext, moduleName: string, source: string): handler.Result {
+					let theModule: componentsModules.Module = registry.getModule(moduleName);
+					if (coreTypes.isNothing(theModule)) {
+						return {
+							status: handler.ResultStatus.Failure,
+							error: `No such module '${moduleName}'`
+						}
+					} else if (coreTypes.isNothing(theModule.getRemote().base(source))) {
+						return {
+							status: handler.ResultStatus.Failure,
+							error: `Module '${moduleName}' has no remote description for '${ source}'`
+						}
+					} else {
+						context.getParent().setRemoteSource(theModule.getPath(), source);
+						return {
+							status: handler.ResultStatus.Success,
+							value: `set '${moduleName}' to work with remote source '${source}'`
+						}
+					}
+				}
+			}
+		}
 	};
 
-	export type ContextProvider = {
-		terminal: () => TerminalContext;
-		module: {
-			restricted: (path: components.Path) => modules.ModuleContext;
-			privileged: (path: components.Path) => modules.PrivilegedModuleContext;
-		}
-	}
-
-	function moduleContextProvider(this: { modules: Map<string, LoadedModule> }, type: "restricted" | "privileged", path: components.Path) {
-		while (path.length > 0) {
-			const key = path.toString();
-
-			if (this.modules.has(key)) {
-				return this.modules.get(key).context[type];
-			}
-
-			path = path.parent();
-		}
-
-		return null;
-	}
-
-	export class Terminal {
-		private properties: Properties;
-		private view: view.TerminalView;
-		private context: TerminalContext;
-		private contextProvider: ContextProvider;
-		private modules: Map<string, LoadedModule>;
-		private variables: collections.Map<Variable>;
-
-		public constructor(properties: Properties, applicationContext: ApplicationContext, viewFactory: view.TerminalFactory) {
-			this.properties = properties;
-			this.modules = new Map();
-			this.variables = collections.map<Variable>();
-			this.context = new TerminalContext(this, applicationContext);
-			this.contextProvider = {
-				terminal: () => this.context,
-				module: {
-					restricted: moduleContextProvider.bind(this, "restricted"),
-					privileged: moduleContextProvider.bind(this, "privileged")
-				}
-			}
-
-			this.moduleLoaded(components.registry.getModule("io.fugazi"));
-
-			viewFactory.createTerminal({
-				name: properties.name,
-				title: properties.title || properties.name,
-				description: properties.description,
-				history: this.properties.history,
-				querier: this.queryForStatements.bind(this),
-				executer: this.executeCommand.bind(this)
-			}).then(this.setView.bind(this));
-		}
-
-		public clearOutput(): void {
-			this.view.clearOutput();
-		}
-
-		public moduleLoaded(module: components.modules.Module) {
-			const path = module.getPath().toString();
-
-			if (!this.modules.has(path)) {
-				this.modules.set(path, {
-					module,
-					context: {
-						restricted: new modules.ModuleContext(this.context.asRestricted()),
-						privileged: new modules.PrivilegedModuleContext(this.context)
-					}
-				});
-			}
-		}
-
-		public storeVariable(name: string, type: components.types.Type, value: any): Variable {
-			let variable = {
-				name: extractVariableName(name),
-				type: type,
-				value: value
-			};
-
-			this.variables.set(variable.name, variable);
-			return variable;
-		}
-
-		public retrieveVariable(name: string): Variable {
-			return this.variables.get(extractVariableName(name));
-		}
-
-		private setView(view: view.TerminalView): void {
-			this.view = view;
-		}
-
-		private executeCommand(command: string): components.commands.ExecutionResult {
-			ga("send", "event", "Commands", "execution - start", command);
-			storage.local.store(this.properties.name, this.properties);
-
-			let result: components.commands.ExecutionResult = null;
-			try {
-				let session = statements.createStatementsSession(command, this.contextProvider),
-					executableStatement = session.getExecutable();
-
-				if (!isNull(executableStatement)) {
-					result = executableStatement.execute();
-				} else {
-					ga("send", "event", "Commands", "execution - None of the statements are executable", command);
-					throw new Exception("None of the statements are executable");
-				}
-			} catch (e) {
-				const error = typeof e === "string" ? e : (e.message ? e.message : e.toString());
-				ga("send", "event", "Commands", "execution - error: " + error, command);
-				result = new components.commands.ExecutionResult(components.registry.getType("any"), false);
-				result.reject(e);
-			}
-
-			return result;
-		}
-		
-		private queryForStatements(command: string, position: number): Promise<app.statements.Statement[]> {
-			let future = new Future<app.statements.Statement[]>(),
-				session = statements.createStatementsSession(command, this.contextProvider),
-				suggestedStatements = session.getSuggestions(position);
-
-			future.resolve(suggestedStatements);
-			
-			return future.asPromise();
-		}
-	}
-
-	function extractVariableName(input: string): string {
-		return input.startsWith("$") ? input.substring(1) : input;
-	}
-
-	// init
-	app.bus.register(app.Events.Ready, function(): void {
-		let coreCommands = {
-			name: "io.fugazi.terminal",
-			title: "Terminal Module",
-			constraints: {
-				variable: {
-					title: "Variable name constraint",
-					types: ["string"],
-					params: [],
-					validator: () => {
-						return function(value: string): boolean {
-							return /^\$[^"':\[]\(\)\{}]+$/.test(value);
-						}
-					}
-				}
-			},
-			types: {
-				"variable": {
-					title: "Variable Token",
-					type: "string[variable]"
-				}
-			},
-			commands: {
-				clear: {
-					title: "Clears the output panel",
-					syntax: "clear",
-					returns: "void",
-					componentConstructor: TerminalCommand,
-					handler: function(context: modules.PrivilegedModuleContext): components.commands.handler.Result {
-						context.getParent().clearOutput();
-						return {
-							status: components.commands.handler.ResultStatus.Success
-						}
-					}
-				},
-				assign: {
-					title: "Variable assignment command",
-					syntax: [
-						"(name string) = (value any)",
-						"set (name string) = (value any)",
-						"let (name string) = (value any)"
-					],
-					returns: "void",
-					parametersForm: "arguments",
-					componentConstructor: TerminalCommand,
-					handler: function(context: modules.PrivilegedModuleContext, name: string, value: any): components.commands.handler.Result {
-						let type = context.guessType(value);
-						if (type == null) {
-							return {
-								status: components.commands.handler.ResultStatus.Failure,
-								error: "can't guess type"
-							}
-						}
-
-						context.getParent().storeVariable(name, type, value);
-						return {
-							status: components.commands.handler.ResultStatus.Success
-						};
-					}
-				},
-				moduleWorkWith: {
-					title: "Specific module remote source setup",
-					syntax: [
-						"module (moduleName string) works with (source string)",
-					],
-					returns: "ui.message",
-					parametersForm: "arguments",
-					componentConstructor: TerminalCommand,
-					handler: function (context: modules.ModuleContext, moduleName: string, source: string): components.commands.handler.Result {
-						let theModule: components.modules.Module = components.registry.getModule(moduleName);
-						if (isNothing(theModule)) {
-							return {
-								status: components.commands.handler.ResultStatus.Failure,
-								error: `No such module '${moduleName}'`
-							}
-						} else if (isNothing(theModule.getRemote().base(source))) {
-							return {
-								status: components.commands.handler.ResultStatus.Failure,
-								error: `Module '${moduleName}' has no remote description for '${ source}'`
-							}
-						} else {
-							context.getParent().setRemoteSource(theModule.getPath(), source);
-							return {
-								status: components.commands.handler.ResultStatus.Success,
-								value: `set '${moduleName}' to work with remote source '${source}'`
-							}
-						}
-					}
-				}
-			}
-		};
-
-		let moduleBuilder = components.modules.builder.create(coreCommands) as components.modules.builder.Builder;
-		moduleBuilder.build().then((aModule: components.modules.Module) => {
-			moduleBuilder.associate();
-			components.registry.add(aModule);
-		}).catch(e => { throw e });
+	let moduleBuilder = modulesBuilder.create(coreCommands) as modulesBuilder.Builder;
+	moduleBuilder.build().then((aModule: componentsModules.Module) => {
+		moduleBuilder.associate();
+		registry.add(aModule);
+	}).catch(e => {
+		throw e
 	});
-}
+});
