+++ scripts/src/components/modules.ts
@@ -18,13 +18,14 @@ import * as syntax from "./syntax";
 import * as constraints from "./constraints";
 import * as converters from "./converters";
 import * as commands from "./commands";
+import * as descriptor from "./modules.descriptor";
 import * as appFrames from "../app/frames";
 import * as appModules from "../app/modules";
 import * as app from "../app/application";
 
-type LookupType = "module-context";
+export type LookupType = "module-context";
 
-type Lookup = {
+export type Lookup = {
 	type: LookupType;
 	name: string;
 };
@@ -39,7 +40,7 @@ function getLookupType(str: string): LookupType {
 }
 
 const LOOKUP_PATTERN = /^(module):([a-zA-Z]+[a-zA-Z0-9._-]*)$/;
-function createLookup(str: string): Lookup {
+export function createLookup(str: string): Lookup {
 	if (!LOOKUP_PATTERN.test(str)) {
 		throw new coreTypes.Exception(`"${ str } not a valid lookup`);
 	}
@@ -51,7 +52,7 @@ function createLookup(str: string): Lookup {
 	};
 }
 
-class Parameters {
+export class Parameters {
 	private lookup: Map<string, Lookup>;
 	private defaults: Map<string, string>;
 
@@ -600,712 +601,6 @@ export class RemotePath implements Remote {
 	}
 }
 
-function isRemotePath(remote: Remote): remote is RemotePath {
+export function isRemotePath(remote: Remote): remote is RemotePath {
 	return remote instanceof RemotePath;
-}
-
-export namespace descriptor {
-	var loadingScriptModules: collections.Map<ScriptLoader> = collections.map<ScriptLoader>();
-
-	/**
-	 * Called by a loaded (script) component
-	 * @param descriptor
-	 */
-	export function loaded(moduleDescriptor: Descriptor): void {
-		let currentScript = document.currentScript as HTMLScriptElement,
-			name: string = currentScript.id;
-
-		if (!loadingScriptModules.has(name)) {
-			logger.error("unexpected module loaded");
-			dom.remove(currentScript);
-			return;
-		}
-
-		loadingScriptModules.remove(name).loadedCalled(moduleDescriptor);
-	}
-
-	export interface InnerComponentsCollection<T extends componentsDescriptor.Descriptor> {
-	}
-
-	export interface InnerComponentsArrayCollection<T extends componentsDescriptor.Descriptor> extends InnerComponentsCollection<T>, Array<string | T> {
-	}
-
-	export interface InnerComponentsObjectCollection<T extends componentsDescriptor.Descriptor> extends InnerComponentsCollection<T>, coreTypes.PlainObject<string | T> {
-	}
-
-	export interface InnerModulesCollection {
-	}
-
-	export interface InnerModulesArrayCollection extends InnerModulesCollection, Array<Descriptor | string> {
-	}
-
-	export interface InnerModulesObjectCollection extends InnerModulesCollection, coreTypes.PlainObject<Descriptor | string> {
-	}
-
-	export interface Descriptor extends componentsDescriptor.Descriptor {
-		basePath?: string; // set when building
-		remote?: RemoteDescriptor;
-		lookup?: { [name: string]: string };
-		modules?: InnerModulesCollection;
-		types?: string | InnerComponentsCollection<types.descriptor.Descriptor>;
-		commands?: string | InnerComponentsCollection<commandsDescriptor.Descriptor>;
-		converters?: string | InnerComponentsCollection<convertersDescriptor.Descriptor>;
-		constraints?: string | InnerComponentsCollection<constraintsDescriptor.Descriptor>;
-	}
-
-	export type AuthenticationMethod = "none" | "basic" | "custom";
-
-	export interface BaseSourceRemoteDescriptor {
-		base?: string;
-		proxy?: string;
-		default?: string;
-		auth?: AuthenticationMethod;
-	}
-
-	export interface SingleSourceRemoteDescriptor extends BaseSourceRemoteDescriptor {
-		origin: string;
-	}
-
-	export interface MultiSourceRemoteDescriptor extends BaseSourceRemoteDescriptor {
-		origins: { [name: string]: string };
-	}
-
-	export type SourceRemoteDescriptor = SingleSourceRemoteDescriptor | MultiSourceRemoteDescriptor;
-
-	export function isSingleSourceRemoteDescriptor(remote: SourceRemoteDescriptor): remote is SingleSourceRemoteDescriptor {
-		return (remote as SingleSourceRemoteDescriptor).origin != null;
-	}
-
-	export function isMultiSourceRemoteDescriptor(remote: SourceRemoteDescriptor): remote is MultiSourceRemoteDescriptor {
-		return (remote as MultiSourceRemoteDescriptor).origins != null;
-	}
-
-	export interface RelativeRemoteDescriptor {
-		path: string;
-	}
-
-	export type RemoteDescriptor = SourceRemoteDescriptor | RelativeRemoteDescriptor;
-
-	export function isRelativeRemoteDescriptor(remote: RemoteDescriptor): remote is RelativeRemoteDescriptor {
-		return (remote as RelativeRemoteDescriptor).path != null;
-	}
-
-	export abstract class RemoteLoader extends componentsDescriptor.BaseLoader<Descriptor> {
-		protected cached: boolean;
-		protected future: coreTypes.Future<Descriptor>;
-
-		constructor(url: net.Url, cached: boolean) {
-			super(url);
-			this.cached = cached;
-			this.future = new coreTypes.Future<Descriptor>();
-		}
-
-		public then(fn: (aDescriptor: Descriptor) => Descriptor): componentsDescriptor.Loader<Descriptor> {
-			this.future = this.future.then<Descriptor>(fn);
-			return this;
-		}
-
-		public catch(fn: (error: coreTypes.Exception) => void): componentsDescriptor.Loader<Descriptor> {
-			this.future = this.future.catch(fn);
-			return this;
-		}
-	}
-
-	export class HttpLoader extends RemoteLoader {
-		constructor(url: net.Url, cached: boolean = true) {
-			super(url, cached);
-
-			url = url.clone();
-			if (!cached) {
-				url.addParam("dummy", new Date().getTime());
-			}
-
-			net.get({url: url, cors: true})
-				.success(this.loaded.bind(this))
-				.fail(this.failed.bind(this))
-				.send();
-		}
-
-		private loaded(response: net.HttpResponse): void {
-			try {
-				this.future.resolve(response.getDataAsJson());
-			} catch (e) {
-				this.future.reject(new coreTypes.Exception("illegal module descriptor"));
-			}
-		}
-
-		private failed(response: net.HttpResponse): void {
-			this.future.reject(new coreTypes.Exception(response.getData()));
-		}
-	}
-
-	export class ScriptLoader extends RemoteLoader {
-		private static LOADED_TIMEOUT: number = 1000;
-
-		private loaded: boolean;
-		private scriptId: string;
-		private timer: utils.Timer;
-
-		constructor(url: net.Url, cached: boolean = true) {
-			super(url, cached);
-
-			this.loaded = false;
-			this.scriptId = utils.generateId({prefix: "module_", min: 6, max: 10});
-			loadingScriptModules.set(this.scriptId, this);
-			utils.loadScript(this.baseUrl.toString(), this.scriptId, this.cached).then(
-				this.scriptLoaded.bind(this),
-				this.failed.bind(this, "component loading timeout"));
-		}
-
-		public loadedCalled(moduleDescriptor: Descriptor): void {
-			this.loaded = true;
-			this.timer && this.timer.cancel();
-			this.future.resolve(moduleDescriptor);
-		}
-
-		private scriptLoaded(): void {
-			if (!this.loaded) {
-				this.timer = new utils.Timer(ScriptLoader.LOADED_TIMEOUT, this.failed.bind(this, "component did not call loaded"));
-			}
-		}
-
-		private failed(message: string): void {
-			this.timer && this.timer.cancel();
-			this.future.reject(new coreTypes.Exception(message));
-		}
-	}
-}
-
-export namespace builder {
-	function breakDescriptorWithPath(moduleDescriptor: descriptor.Descriptor) {
-		let result: descriptor.Descriptor,
-			path: string[] = moduleDescriptor.name.split(".");
-
-		if (path.length === 1) {
-			result = moduleDescriptor;
-		} else {
-			let first = path.remove(0),
-				tail = path.join("."),
-				innerModuleDescriptor = {} as descriptor.Descriptor;
-
-			Object.keys(moduleDescriptor).forEach(key => innerModuleDescriptor[key] = moduleDescriptor[key]);
-			innerModuleDescriptor.name = tail;
-
-			result = {
-				name: first,
-				modules: [breakDescriptorWithPath(innerModuleDescriptor)]
-			}
-		}
-
-		return result;
-	}
-
-	export function create(url: net.Url): componentsBuilder.Builder<Module>;
-	export function create(moduleDescriptor: descriptor.Descriptor, parent?: componentsBuilder.Builder<components.Component>): componentsBuilder.Builder<Module>;
-	export function create(moduleDescriptor: net.Url | descriptor.Descriptor, parent?):componentsBuilder.Builder<Module> {
-		let loader: componentsDescriptor.Loader<descriptor.Descriptor>;
-
-		if (moduleDescriptor instanceof net.Url) {
-			if (moduleDescriptor.pathname.endsWith(".js")) {
-				loader = new descriptor.ScriptLoader(moduleDescriptor);
-			} else if (moduleDescriptor.pathname.endsWith(".json")) {
-				loader = new descriptor.HttpLoader(moduleDescriptor);
-			} else {
-				throw new coreTypes.Exception("can only load .js or .json urls");
-			}
-		} else {
-			if (parent) {
-				loader = new componentsDescriptor.ExistingLoader(moduleDescriptor, (<any> parent).loader.getUrl());
-			} else {
-				loader = new componentsDescriptor.ExistingLoader(moduleDescriptor);
-			}
-		}
-
-		loader.then(aDescriptor => {
-			const tree = breakDescriptorWithPath(aDescriptor);
-
-			if (!parent) {
-				(tree as any).basePath = aDescriptor.name;
-			}
-
-			return tree;
-		});
-		return new Builder(loader, parent);
-	}
-
-	interface RemoteBuilderInfo {
-		path: components.Path;
-		builder: Builder;
-		name?: string;
-	}
-
-	export class Builder extends componentsBuilder.BaseBuilder<Module, descriptor.Descriptor> {
-		private innerModuleBuilders: collections.Map<componentsBuilder.Builder<Module>>;
-		private innerTypesBuilders: collections.Map<componentsBuilder.Builder<types.Type>>;
-		private innerCommandsBuilders: collections.Map<componentsBuilder.Builder<commands.Command>>;
-		private innerConvertersBuilders: collections.Map<componentsBuilder.Builder<converters.Converter>>;
-		private innerConstraintBuilders: collections.Map<componentsBuilder.Builder<constraints.Constraint>>;
-
-		private innerModuleRemoteBuilders: Array<RemoteBuilderInfo>;
-		private innerTypeModuleRemoteBuilders: Array<RemoteBuilderInfo>;
-		private innerCommandModuleRemoteBuilders: Array<RemoteBuilderInfo>;
-		private innerConvertersModuleRemoteBuilders: Array<RemoteBuilderInfo>;
-		private innerConstraintModuleRemoteBuilders: Array<RemoteBuilderInfo>;
-
-		private basePath: components.Path;
-		private remote: Remote;
-		private params: Parameters;
-
-		public constructor(loader: componentsDescriptor.Loader<descriptor.Descriptor>, parent?: componentsBuilder.Builder<components.Component>) {
-			super(Module, loader, parent);
-		}
-
-		public resolve<C2 extends components.Component>(type: components.ComponentType, name: string): C2 {
-			var path: string[] = name.split("."),
-				collection: collections.Map<componentsBuilder.Builder<components.Component>> = null;
-
-			if (path.length > 1) {
-				if (this.getName() === path.first()) {
-					name = name.substring(name.indexOf(".") + 1);
-				} else if (this.innerModuleBuilders.has(path.first())) {
-					return this.innerModuleBuilders.get(path.first()).resolve<C2>(type, name.substring(name.indexOf(".") + 1));
-				}
-			}
-
-			switch (type) {
-				case components.ComponentType.Type:
-					collection = this.innerTypesBuilders;
-					break;
-
-				case components.ComponentType.Module:
-					collection = this.innerModuleBuilders;
-					break;
-
-				case components.ComponentType.Command:
-					collection = this.innerCommandsBuilders;
-					break;
-
-				case components.ComponentType.Converter:
-					collection = this.innerConvertersBuilders;
-					break;
-
-				case components.ComponentType.Constraint:
-					collection = this.innerConstraintBuilders;
-					break;
-			}
-
-			if (collection !== null && collection.has(name)) {
-				return <C2> collection.get(name).getComponent();
-			}
-
-			return super.resolve<C2>(type, name);
-		}
-
-		public getBasePath() {
-			return this.getParent() ? (this.getParent() as Builder).getBasePath() : this.basePath;
-		}
-
-		protected onDescriptorReady(): void {
-			this.createParameters();
-
-			this.innerModuleRemoteBuilders = [];
-			this.innerTypeModuleRemoteBuilders = [];
-			this.innerCommandModuleRemoteBuilders = [];
-			this.innerConvertersModuleRemoteBuilders = [];
-			this.innerConstraintModuleRemoteBuilders = [];
-
-			this.innerModuleBuilders = collections.map<componentsBuilder.Builder<Module>>();
-			this.innerTypesBuilders = collections.map<componentsBuilder.Builder<types.Type>>();
-			this.innerCommandsBuilders = collections.map<componentsBuilder.Builder<commands.Command>>();
-			this.innerConvertersBuilders = collections.map<componentsBuilder.Builder<converters.Converter>>();
-			this.innerConstraintBuilders = collections.map<componentsBuilder.Builder<constraints.Constraint>>();
-
-			if (!coreTypes.isNothing(this.componentDescriptor.modules)) {
-				this.createInnerModuleBuilders();
-			}
-
-			if (!coreTypes.isNothing(this.componentDescriptor.constraints)) {
-				this.createInnerConstraintBuilders();
-			}
-
-			if (!coreTypes.isNothing(this.componentDescriptor.types)) {
-				this.createInnerTypeBuilders();
-			}
-
-			if (!coreTypes.isNothing(this.componentDescriptor.commands)) {
-				this.createInnerCommandBuilders();
-			}
-
-			if (!coreTypes.isNothing(this.componentDescriptor.converters)) {
-				this.createInnerConvertersBuilders();
-			}
-
-			if ((<descriptor.Descriptor> this.componentDescriptor).remote) {
-				const remote = <descriptor.RemoteDescriptor>(<descriptor.Descriptor> this.componentDescriptor).remote;
-
-				this.handleRemoteDescriptor(remote);
-			}
-
-			if (!this.getParent()) {
-				this.basePath = new components.Path(this.componentDescriptor.basePath);
-			}
-		}
-
-		protected concreteBuild(): void {
-			let innerBuilders = <Array<componentsBuilder.Builder<components.Component>>> [].concat(this.innerModuleBuilders.values())
-				.concat(this.innerCommandsBuilders.values())
-				.concat(this.innerConstraintBuilders.values())
-				.concat(this.innerTypesBuilders.values())
-				.concat(this.innerConvertersBuilders.values());
-
-			if (innerBuilders.length > 0) {
-				innerBuilders.forEach(componentBuilder => {
-					componentBuilder.build().then(this.innerBuilderCompleted.bind(this), this.future.reject);
-					this.innerBuilderCreated();
-				});
-			} else if (this.component) {
-				this.future.resolve(this.component);
-			}
-
-			this.innerModuleRemoteBuilders.forEach(pair => {
-				pair.builder.build().then(this.remoteModuleBuilt.bind(this, pair, "module"), this.future.reject);
-				this.innerBuilderCreated();
-			});
-
-			this.innerTypeModuleRemoteBuilders.forEach(pair => {
-				pair.builder.build().then(this.remoteModuleBuilt.bind(this, pair, "type"), this.future.reject);
-				this.innerBuilderCreated();
-			});
-
-			this.innerCommandModuleRemoteBuilders.forEach(pair => {
-				pair.builder.build().then(this.remoteModuleBuilt.bind(this, pair, "command"), this.future.reject);
-				this.innerBuilderCreated();
-			});
-
-			this.innerConvertersModuleRemoteBuilders.forEach(pair => {
-				pair.builder.build().then(this.remoteModuleBuilt.bind(this, pair, "converter"), this.future.reject);
-				this.innerBuilderCreated();
-			});
-
-			this.innerConstraintModuleRemoteBuilders.forEach(pair => {
-				pair.builder.build().then(this.remoteModuleBuilt.bind(this, pair, "constraint"), this.future.reject);
-				this.innerBuilderCreated();
-			});
-		}
-
-		protected concreteAssociate(): void {
-			(this.component as any).remote = this.remote;
-			(this.component as any).params = this.params;
-
-			this.innerModuleBuilders.forEach((moduleBuilder: Builder) => {
-				if (isRemotePath(moduleBuilder.remote)) {
-					(moduleBuilder.remote as any)._parent = this.remote;
-				}
-				moduleBuilder.associate();
-				(<any> this).component.modules.set(moduleBuilder.getName(), moduleBuilder.getComponent());
-			});
-
-			this.innerConstraintBuilders.forEach(function (constraintBuilder: componentsBuilder.Builder<constraints.Constraint>): void {
-				constraintBuilder.associate();
-				(<any> this).component.constraints.set(constraintBuilder.getName(), constraintBuilder.getComponent());
-			}.bind(this));
-
-			this.innerTypesBuilders.forEach(function (typeBuilder: componentsBuilder.Builder<types.Type>): void {
-				typeBuilder.associate();
-				(<any> this).component.types.set(typeBuilder.getName(), typeBuilder.getComponent());
-			}.bind(this));
-
-			this.innerCommandsBuilders.forEach(function (commandBuilder: componentsBuilder.Builder<commands.Command>): void {
-				commandBuilder.associate();
-				(<any> this).component.commands.set(commandBuilder.getName(), commandBuilder.getComponent());
-			}.bind(this));
-
-			this.innerConvertersBuilders.forEach(function (converterBuilder: componentsBuilder.Builder<converters.Converter>): void {
-				converterBuilder.associate();
-				(<any> this).component.converters.set(converterBuilder.getName(), converterBuilder.getComponent());
-			}.bind(this));
-		}
-
-		private createParameters() {
-			this.params = new Parameters();
-
-			if (this.componentDescriptor.lookup) {
-				Object.keys(this.componentDescriptor.lookup).forEach(name => {
-					(this.params as any).lookup.set(name, createLookup(this.componentDescriptor.lookup[name]));
-				});
-			}
-		}
-
-		private createInnerModuleBuilders() {
-			if (coreTypes.is(this.componentDescriptor.modules, Array)) {
-				(<descriptor.InnerModulesArrayCollection> this.componentDescriptor.modules).forEach(current => {
-					if (typeof current === "string") { // url
-						this.innerModuleRemoteBuilders.push({
-							path: this.getPath().clone(),
-							builder: <Builder> create(this.createInnerUrl(current))
-						});
-					} else { // descriptor
-						this.innerModuleBuilders.set(current.name, create(current, this));
-					}
-				});
-			} else if (coreTypes.isPlainObject(this.componentDescriptor.modules)) {
-				Object.keys(this.componentDescriptor.modules).forEach(moduleName => {
-					let moduleDescriptor = this.componentDescriptor.modules[moduleName];
-
-					if (typeof moduleDescriptor === "string") { // url
-						this.innerModuleRemoteBuilders.push({
-							name: moduleName,
-							path: this.getPath().clone(),
-							builder: <Builder> create(this.createInnerUrl(moduleDescriptor))
-						});
-					} else { // descriptor
-						moduleDescriptor.name = moduleName;
-						this.innerModuleBuilders.set(moduleName, create(moduleDescriptor, this));
-					}
-				});
-			} else {
-				throw new componentsBuilder.Exception("can only build inner modules from array or object collections");
-			}
-		}
-
-		private createInnerCommandBuilders() {
-			if (typeof this.componentDescriptor.commands === "string") { // url
-				this.innerCommandModuleRemoteBuilders.push({
-					path: this.getPath().clone(),
-					builder: <Builder> create(this.createInnerUrl(<string> this.componentDescriptor.commands))
-				});
-			} else if (this.componentDescriptor.commands instanceof Array) {
-				(<descriptor.InnerComponentsArrayCollection<commandsDescriptor.Descriptor>> this.componentDescriptor.commands).forEach(current => {
-					if (typeof current === "string") { // url
-						this.innerCommandModuleRemoteBuilders.push({
-							path: this.getPath().clone(),
-							builder: <Builder> create(this.createInnerUrl(current))
-						});
-					} else { // descriptor
-						this.innerCommandsBuilders.set(current.name, commandsBuilder.create(current, this));
-					}
-				});
-			} else if (coreTypes.isPlainObject(this.componentDescriptor.commands)) {
-				Object.keys(this.componentDescriptor.commands).forEach(commandName => {
-					let commandDescriptor = this.componentDescriptor.commands[commandName];
-
-					if (typeof commandDescriptor === "string") { // url
-						this.innerCommandModuleRemoteBuilders.push({
-							name: commandName,
-							path: this.getPath().clone(),
-							builder: <Builder> create(this.createInnerUrl(commandDescriptor))
-						});
-					} else { // descriptor
-						commandDescriptor.name = commandName;
-						this.innerCommandsBuilders.set(commandName, commandsBuilder.create(commandDescriptor, this));
-					}
-				});
-			} else {
-				throw new componentsBuilder.Exception("can only build commands from url string, array or object collections");
-			}
-		}
-
-		private createInnerConvertersBuilders() {
-			if (typeof this.componentDescriptor.converters === "string") { // url
-				this.innerConvertersModuleRemoteBuilders.push({
-					path: this.getPath().clone(),
-					builder: <Builder> create(this.createInnerUrl(<string> this.componentDescriptor.converters))
-				});
-			} else if (this.componentDescriptor.converters instanceof Array) {
-				(<descriptor.InnerComponentsArrayCollection<convertersDescriptor.Descriptor>> this.componentDescriptor.converters).forEach(current => {
-					if (typeof current === "string") { // url
-						this.innerConvertersModuleRemoteBuilders.push({
-							path: this.getPath().clone(),
-							builder: <Builder> create(this.createInnerUrl(current))
-						});
-					} else { // descriptor
-						this.innerConvertersBuilders.set(current.name, convertersBuilder.create(current, this));
-					}
-				});
-			} else if (coreTypes.isPlainObject(this.componentDescriptor.converters)) {
-				Object.keys(this.componentDescriptor.converters).forEach(converterName => {
-					let converterDescriptor = this.componentDescriptor.converters[converterName];
-
-					if (typeof converterDescriptor === "string") { // url
-						this.innerConvertersModuleRemoteBuilders.push({
-							name: converterName,
-							path: this.getPath().clone(),
-							builder: <Builder> create(this.createInnerUrl(converterDescriptor))
-						});
-					} else { // descriptor
-						converterDescriptor.name = converterName;
-						this.innerConvertersBuilders.set(converterName, convertersBuilder.create(converterDescriptor, this));
-					}
-				});
-			} else {
-				throw new componentsBuilder.Exception("can only build converters from url string, array or object collections");
-			}
-		}
-
-		private createInnerConstraintBuilders() {
-			if (typeof this.componentDescriptor.constraints === "string") { // url
-				this.innerConstraintModuleRemoteBuilders.push({
-					path: this.getPath().clone(),
-					builder: <Builder> create(this.createInnerUrl(<string> this.componentDescriptor.constraints))
-				});
-			} else if (this.componentDescriptor.constraints instanceof Array) {
-				(<descriptor.InnerComponentsArrayCollection<constraintsDescriptor.Descriptor>> this.componentDescriptor.constraints).forEach(current => {
-					if (typeof current === "string") { // url
-						this.innerConstraintModuleRemoteBuilders.push({
-							path: this.getPath().clone(),
-							builder: <Builder> create(this.createInnerUrl(current))
-						});
-					} else { // descriptor
-						this.innerConstraintBuilders.set(current.name, constraintsBuilder.create(current, this));
-					}
-				});
-			} else if (coreTypes.isPlainObject(this.componentDescriptor.constraints)) {
-				Object.keys(this.componentDescriptor.constraints).forEach(constraintName => {
-					let constraintDescriptor = this.componentDescriptor.constraints[constraintName];
-
-					if (typeof constraintDescriptor === "string") { // url
-						this.innerConstraintModuleRemoteBuilders.push({
-							name: constraintName,
-							path: this.getPath().clone(),
-							builder: <Builder> create(this.createInnerUrl(constraintDescriptor))
-						});
-					} else { // descriptor
-						constraintDescriptor.name = constraintName;
-						this.innerConstraintBuilders.set(constraintName, constraintsBuilder.create(constraintDescriptor, this));
-					}
-				});
-			} else {
-				throw new componentsBuilder.Exception("can only build constraints from url string, array or object collections");
-			}
-		}
-
-		private createInnerTypeBuilders() {
-			if (typeof this.componentDescriptor.types === "string") { // url
-				this.innerTypeModuleRemoteBuilders.push({
-					path: this.getPath().clone(),
-					builder: <Builder> create(this.createInnerUrl(<string> this.componentDescriptor.types))
-				});
-			} else if (this.componentDescriptor.types instanceof Array) {
-				(<descriptor.InnerComponentsArrayCollection<types.descriptor.Descriptor>> this.componentDescriptor.types).forEach(current => {
-					if (typeof current === "string") { // url
-						this.innerTypeModuleRemoteBuilders.push({
-							path: this.getPath().clone(),
-							builder: <Builder> create(this.createInnerUrl(current))
-						});
-					} else { // descriptor
-						this.innerTypesBuilders.set(current.name, types.builder.create(current, this));
-					}
-				});
-			} else if (coreTypes.isPlainObject(this.componentDescriptor.types)) {
-				Object.keys(this.componentDescriptor.types).forEach(typeName => {
-					let typeDescriptor = this.componentDescriptor.types[typeName];
-
-					if (typeof typeDescriptor === "string") { // url
-						this.innerTypeModuleRemoteBuilders.push({
-							name: typeName,
-							path: this.getPath().clone(),
-							builder: <Builder> create(this.createInnerUrl(typeDescriptor))
-						});
-					} else { // descriptor
-						typeDescriptor.name = typeName;
-						this.innerTypesBuilders.set(typeName, types.builder.create(typeDescriptor, this));
-					}
-				});
-			} else {
-				throw new componentsBuilder.Exception("can only build constraints from url string, array or object collections");
-			}
-		}
-
-		private createInnerUrl(path: string): net.Url {
-			return this.loader.getUrlFor(path);
-		}
-
-		private remoteModuleBuilt(info: RemoteBuilderInfo, collectionType: components.ComponentTypeName): void {
-			let newBuilder = info.builder;
-
-			for (let i = 1; newBuilder != null && i < info.path.length; i++) {
-				newBuilder = newBuilder.innerModuleBuilders.values()[0] as Builder;
-			}
-
-			if (newBuilder == null) {
-				throw new componentsBuilder.Exception(`remote module path doesn't match parent path 1`);
-			}
-
-			switch (collectionType) {
-				case "module":
-					newBuilder = <Builder> newBuilder.innerModuleBuilders.values()[0];
-					if (newBuilder == null) {
-						throw new componentsBuilder.Exception(`remote module path doesn't match parent path 2`);
-					}
-
-					newBuilder.parent = this;
-					this.innerModuleBuilders.set(info.name || info.builder.getName(), newBuilder);
-
-					break;
-
-				case "command":
-					newBuilder.innerCommandsBuilders.values().forEach(commandBuilder => {
-						(<any> commandBuilder).parent = this;
-						this.innerCommandsBuilders.set(commandBuilder.getName(), commandBuilder);
-					});
-
-					break;
-
-				case "type":
-					newBuilder.innerTypesBuilders.values().forEach(typeBuilder => {
-						(<any> typeBuilder).parent = this;
-						this.innerTypesBuilders.set(typeBuilder.getName(), typeBuilder);
-					});
-
-					break;
-
-				case "converter":
-					newBuilder.innerConvertersBuilders.values().forEach(converterBuilder => {
-						(<any> converterBuilder).parent = this;
-						this.innerConvertersBuilders.set(converterBuilder.getName(), converterBuilder);
-					});
-
-					break;
-
-				case "constraint":
-					newBuilder.innerConstraintBuilders.values().forEach(constraintBuilder => {
-						(<any> constraintBuilder).parent = this;
-						this.innerConstraintBuilders.set(constraintBuilder.getName(), constraintBuilder);
-					});
-
-					break;
-			}
-
-			this.innerBuilderCompleted();
-		}
-
-		private handleRemoteDescriptor(remote: descriptor.RemoteDescriptor) {
-			if (descriptor.isRelativeRemoteDescriptor(remote)) {
-				this.handleRemoteRelativeDescriptor(remote);
-			} else {
-				this.handleRemoteSourceDescriptor(remote);
-			}
-		}
-
-		private handleRemoteRelativeDescriptor(relative: descriptor.RelativeRemoteDescriptor) {
-			this.remote = new RemotePath(relative.path);
-		}
-
-		private handleRemoteSourceDescriptor(source: descriptor.SourceRemoteDescriptor) {
-			if (this.getParent() != null &&
-				this.getParent().getComponent() != null &&
-				(<Module> this.getParent().getComponent()).isRemote()) {
-
-				throw new componentsBuilder.Exception(
-					`Contradicting 'remote' descriptions been detected on '${this.getPath()}' 
-						and its parent module '${this.getParent().getPath()}'`);
-			}
-
-			this.remote = new RemoteSource(source);
-			const loginCommandDescriptor = (this.remote as RemoteSource).loginCommandDescriptor();
-			if (loginCommandDescriptor != null && Object.keys(loginCommandDescriptor).length > 0) {
-				this.innerCommandsBuilders.set(loginCommandDescriptor.name, commandsBuilder.create(loginCommandDescriptor, this));
-			}
-		}
-	}
 }
